svsm_request: Move VMSA list handling to vmsa_list.rs
No functional change intended.
